requirements.txt
Flask==0.10.1
docker-py==0.2.3
redis==2.7.6
requests==1.2.3
flask-babel
flask-mail
flask-sqlalchemy
flask-user
psycopg2
